requirements.txt
django>=2.2.6
djangorestframework
django-cors-headers >= 3.1.1
django-rest-knox